+++ sourcespec/ssp_output.py
@@ -309,6 +309,8 @@ def _dict2yaml(dict_like, level=0):
     indent = ' ' * 2 * level
     # use oneliners for dict-like objects containing value and uncertainty keys
     if set(target_dict.keys()).intersection(set(value_uncertainty_keys)):
+        # make sure all values are represented as strings
+        target_dict = {key: str(value) for key, value in target_dict.items()}
         oneliner = str(target_dict).replace("'", "")
         return f'{indent}{oneliner}\n'
     lines = ''